// src/modules/upload/interfaces/upload.interface.ts
export interface IUpload {
  IDTURNO: number;
  IDSERVICIO: number;
  NUMTURNO: string;
  REGION: string;
  FECHACREACION: string;
  OFICINA: string;
  SALA: string;
  CLIENTE: string;
  TIPOCLIENTE: string;
  PROCESO: string;
  SUBPROCESO: string;
  AGRUPAMIENTO: string;
  TRAMITE: string;
  COLA: string;
  SECUENCIA: number;
  ESTRANSFERIDO: string;
  TURNO: string;
  ADICI: string;
  REPETICIONES: number;
  ANIOMES: string;
  HORASOLICITUD: string;
  HORAFINESPERA: string;
  HORALLAMADO: string;
  HORAFINLLAMADO: string;
  HORAFINATENCION: string;
  ESPERA: string;
  LLAMADO: string;
  ATENCION: string;
  TOTAL: string;
  USUARIO: string;
  NOMBREUSUARIO: string;
  TERMINAL: string;
  ESTADO: string;
  DIA: string;
  HORA: string;
  NOMBRECLIENTE: string;
  RAZONSOCIAL: string;
  IDENTIFICACION: string;
  TIPOIDENTIFICACION: string;
  IDTURNO_S: number;
  IDSERVICIO_S: number;
  '(SER)Inscripción RUT': string;
  '(SER)Número Formulario': string;
  '(SER)Cantidad Folios': string;
  '(SER)Resultado del Tramite': string;
  '(SER)Gestion del Caso a PST': string;
  '(SER)Observaciones': string;
  '(SER)Tema de Capacitación/Orie': string;
  '(SER)Otros  Servicios': string;
  '(SER)Actualización RUT': string;
  '(SER)Levantamiento Suspensión': string;
  '(SER)Objeto de Campaña': string;
  '(SER)Resultado Cobranzas': string;
  '(SER)Mensaje de Respuesta': string;
  IDTURNO_T: number;
  '(TUR)Tipo de Identificación Tr': string;
  '(TUR)Clasificacion Tramitante': string;
}